+++ mutwo/isis_converters/__init__.py
@@ -1,2 +1,4 @@
+from . import configurations
 from . import constants
+
 from .isis import *