+++ mutwo/isis_converters/constants/__init__.py
@@ -1,4 +1,4 @@
-"""Constants to be used for and with :mod:`mutwo.ext.converters.frontends.isis`.
+"""Constants to be used for and with :mod:`mutwo.isis_converters`.
 
 The file mostly contains different flags for running ISiS.
 The flag definitions are documented
@@ -6,11 +6,6 @@ The flag definitions are documented
 """
 
 from . import XSAMPA
-
-ISIS_PATH = "isis.sh"
-"""The path to the ISiS shell script. When installing ISiS with the packed
-'Install_ISiS_commandline.sh' script, the path should be 'isis.sh'."""
-
 
 SILENT_FLAG = "--quiet"
 """Flag for preventing ISiS from printing any information